refactor: 블록 문자열을 블록 객체로 수정
