+++ src/main/kotlin/minesweeper/README.md
@@ -13,5 +13,5 @@
 
 - [x] 높이와 너비를 반영해 판을 생성할 수 있다.
 - [x] 지뢰의 개수를 반영해 판을 생성할 수 있다.
-- [ ] 일반 블록 객체 만들기
-- [ ] 지뢰 블록 객체 만들기
+- [x] 일반 블록 객체 만들기
+- [x] 지뢰 블록 객체 만들기
