+++ src/main/kotlin/minesweeper/domain/Block.kt
@@ -1,0 +1,7 @@
+package minesweeper.domain
+
+abstract class Block {
+
+    abstract val mark: String
+    abstract fun open()
+}
