+++ src/test/kotlin/minesweeper/MineSweeperBoardTest.kt
@@ -1,5 +1,6 @@
 package minesweeper
 
+import minesweeper.domain.MineSweeperBoard
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
